4th

+++ pages/Data_loading_and_analysis.py
@@ -36,7 +36,7 @@ with st.echo(code_location='below'):
     """### Выгрузим из базы данных таблицу УИК и пользователями деливери"""
 
     vybory_df =get_data('''select uik, avg (delivery2_price_client_rub) as avg_spend
-                            , avg(share) as umg_share 
+                            , avg(share::numeric) as umg_share 
                             from(select * from delivery a
                                 left join vybory_ful b using (uik)
                                 where b.umg_flg::numeric =1) as a
